python-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
redis==5.0.1
docker==7.0.0
kubernetes==29.0.0
python-multipart==0.0.6
pyyaml==6.0.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
motor==3.3.2
python-dateutil==2.8.2
requests==2.31.0
